backend/ingest/ingest_requirements.txt
boto3
pymupdf
google-genai
pinecone
neo4j
python-dotenv